fix type definition
